+++ spec/spec_helper.rb
@@ -19,4 +19,6 @@ RSpec.configure do |config|
   config.define_derived_metadata do |meta|
     meta[:aggregate_failures] = true
   end
+
+  config.include(RuboCop::RSpec::ExpectOffense)
 end